refactor: Titleコンポーネントのディレクトリ修正
