+++ src/components/Group/Body.js
@@ -8,7 +8,7 @@ import {GroupsAPI} from '../../APILink';
 import {CreateGroupModal} from './CreateGroupModal';
 
 import {SelectPerPage} from '../SelectPerPage';
-import {PageTitle} from '../Title';
+import {PageTitle} from '../Utilities/Title';
 import {PrimaryButton} from '../Buttons/PrimaryButton';
 import {AddButtonList} from '../Buttons/Lists/AddButtonList';
 import {InfoCardList} from '../Cards/Lists/InfoCardList';
