+++ src/components/TeachingMaterials/TeachingMaterials.js
@@ -5,7 +5,7 @@ import TeachingMaterialInfo from './TeachingMaterialInfo';
 import Pagination from '../Pagination';
 import {CreateTeachingMaterialModal} from './CreateTeachingMaterialModal';
 import {SelectPerPage} from '../SelectPerPage';
-import {PageTitle} from '../Title';
+import {PageTitle} from '../Utilities/Title';
 import {InfoCardList} from '../Cards/Lists/InfoCardList';
 import {PrimaryButton} from '../Buttons/PrimaryButton';
 import {AddButtonList} from '../Buttons/Lists/AddButtonList';
